mcp_server/requirements.txt
fastmcp
pygame